dts: pm660l: disable CABC
The Mi A2's SDM660 setup has CABC (Content Adaptive Backlight Control)
on the LCD display's backlight for power saving purposes. This is also
known as Dynamic Contrast, as seen on Wikipedia: https://en.wikipedia.org/wiki/Contrast_ratio#Dynamic_contrast_(DC)

Basically, this feature dims the backlight when there is darker content
on the screen, and brightens it when the content gets brighter. While it
may save power for darker images, it can be annoying to look at.

This commit disables the feature.

Signed-off-by: kdrag0n <dragon@khronodragon.com>