hex_defconfig: remove some IO schedulers
THIS COMMIT IS SUBJECT TO CHANGE.

For solid state devices such as Android phones, we need minimal logic in
the IO scheduler. Spending too much time processing logic for IO will
create a bottleneck, so we can eliminate the larger drivers. I decided
to keep Axiety only because latency matters most in a mobile phone, and
benchmarks prove that Anxiety performs the best in this region.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>