hex_defconfig: Reduce maximum block requests
Multiple sources suggest that reducing the maximum allowed block
requests from 128 to 16 can greatly reduce system hangs at the cost of
marginally reduced throughput. The logic here is that when the maximum
block request limit is large, the kernel tends to wait for the pending
requests to finish before continuing to the next group. Throughput is
marginally affected (as previously stated).

https://github.com/nathanchance/bullhead/commit/236dd90ba41d03c6f016d49511887357ec28acaf#diff-809b3e9c83514697076510cb1c1fbc73

In my tests using the Androbench benchmark (with the same scenarios and
the same boot), I recieved these scores:

		512	128	32	16	8	1
---------------------------------------------------------------
Seq Read	284.02	283.11	290.52	296.43	297.63	289.29
Seq Write	175.96	176.1	171.59	225.83	226.14	181.7
Rand Read	34.21	34.9	33.86	34.09	34.99	32.09
Rand Write	9.94	8.33	10.15	8.76	10.13	8.87
SQL Insert	899.57	844.45	1113.25	1134.96	1079.84	1060.24
SQL Update	2390.22	2492.32	2381.12	2486.56	2417.45	2369.91
SQL Delete	2127.6	2011.4	2021.36	2004.01	2083.07	1904.89
---------------------------------------------------------------
Sum		5921.52	5850.61	6021.85	6190.64	6149.25	5846.99
---------------------------------------------------------------

It is clear that using 16 as the maximum number of block requests is
ideal for our device.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>