sched/tune: Request a governor frequency update on boosts
A popular concern with Dynamic SchedTune boosting is that the governor
will be too slow to respond to the boost, as it must wait for the next
time the scheduler requests a frequency update. This commit forces the
governor to recalculate its frequency after each SchedTune boost.
Additionally, we ignore schedutil's rate limits to force frequency
recalculation regardless of the elapsed time.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>