sched/tune: Do not do a Dynamic SchedTune boost if boost is zero
If we don't do this check in the wrapper function, a mutex lock will be
acquired before a value check is done. We can save some time by doing a
zero check here, since 0 will never be larger than the current boost
value (unless we are handling negative SchedTune boost values, but
that's not a concern here, as we don't support negative integer values
here).

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>